capire perchè dia sempre 0 come output

+++ src/components/input.rs
@@ -9,7 +9,7 @@ pub struct Input {
     spikes: Vec<i8>,
     senders: Vec<SyncSender<i8>>,
     // used just for debugging 
-    ts: i8,
+    ts: i32,
 }
 
 impl Input {
