+++ src/components/neural_layer.rs
@@ -49,7 +49,7 @@ impl NeuralLayer {
         return tids;
     }
 
-    pub fn add_synapse(&mut self, neuron: usize, weight: f32, channel: Receiver<i8>) {
+    pub fn add_synapse(&mut self, neuron: usize, weight: f64, channel: Receiver<i8>) {
         // aggiunge una sinapsi ricevendo peso e receiver a un neurone, return di result se neuron è out of bounds
         let s = Synapse::new(weight, channel);
         // println!("adding synapses to neuron [{}]", &neuron);
